make it better

+++ dataDelete/deleteAdminComment.js
@@ -9,14 +9,18 @@ setTimeout(function run() {
 }, 100);
 function deleteAdminComment(){
     app.post('/deleteAdminComment', type, middleware, (req, res) => {
-        base.collection('comments').findOneAndDelete({
-            id : req.body.id,
-        },(err,result)=>{
-            if(err)
-                return console.log(err);
-            //res.redirect('/');
-        });
-        res.json({status:'ok'});
+        if (req.user.role == "admin") {
+            base.collection('comments').findOneAndDelete({
+                id : req.body.id,
+            },(err,result)=>{
+                if(err)
+                    return console.log(err);
+                //res.redirect('/');
+            });
+            res.json({status:'ok'});
+        } else {
+            return res.json({status: "error"})
+        }
     });
 }
 module.exports.deleteAdminComment = deleteAdminComment;
