+++ dataDelete/deleteComment.js
@@ -9,15 +9,19 @@ setTimeout(function run() {
 }, 100);
 function deleteComment(){
     app.post('/deleteComment', type, middleware, (req, res) => {
-        base.collection('comments').findOneAndDelete({
-            id : req.body.id,
-            email: req.user.email
-        },(err,result)=>{
-            if(err)
-                return console.log(err);
-            //res.redirect('/');
-        });
-        res.json({status:'ok'});
+        if (req.user.role == "user") {
+            base.collection('comments').findOneAndDelete({
+                id : req.body.id,
+                email: req.user.email
+            },(err,result)=>{
+                if(err)
+                    return console.log(err);
+                //res.redirect('/');
+            });
+            res.json({status:'ok'});
+        } else {
+            return res.json({status: "error"})
+        }
     });
 }
 module.exports.deleteComment = deleteComment;
